[Update] Added Font Rendering + Fixed a bug - pre-allocation in Renderer
- Added Font Rendering and Font Assets with the respective serialization stuff
- Fixed a big issue regarding pre-allocating vectors in the GeometryPass, ShadowMap and VoxelizationPass where we have to add into a buffer the meshes that are about to be rendered, so instead of push_back, we pre-allocate everything.
- Other small bug fixes

+++ Frost/src/Frost/Platform/Vulkan/SceneRenderPasses/VulkanGeometryPass.h
@@ -37,15 +37,15 @@ namespace Frost
 	{
 		NewIndirectMeshData() = default;
 
-		uint32_t CmdOffset;          /// DONE
+		uint64_t CmdOffset;          /// DONE
 
-		uint32_t TotalMeshOffset;    /// DONE // Same as `TotalSubmeshCount` however with the added Offset from the previous mesh
-		uint32_t SubmeshCount;       /// DONE   // Reponsible for Submesh Count (without instances) and for the Indirect Command Numbers
-		uint32_t InstanceCount;      /// DONE
-		uint32_t TotalSubmeshCount;  /// DONE // SubmeshCount * InstanceCount
+		uint64_t TotalMeshOffset;    /// DONE // Same as `TotalSubmeshCount` however with the added Offset from the previous mesh
+		uint64_t SubmeshCount;       /// DONE   // Reponsible for Submesh Count (without instances) and for the Indirect Command Numbers
+		uint64_t InstanceCount;      /// DONE
+		uint64_t TotalSubmeshCount;  /// DONE // SubmeshCount * InstanceCount
 
-		uint32_t MaterialCount;	     /// DONE
-		uint32_t MaterialOffset;     /// DONE
+		uint64_t MaterialCount;	     /// DONE
+		uint64_t MaterialOffset;     /// DONE
 
 		AssetHandle MeshAssetHandle; /// DONE
 	};
